Update riding for aquatic mobs

+++ common/src/main/java/com/fossil/fossil/entity/prehistoric/Crassigyrinus.java
@@ -77,8 +77,8 @@ public class Crassigyrinus extends PrehistoricSwimming {
     }
 
     @Override
-    public double swimSpeed() {
-        return 0.5;
+    public float swimSpeed() {
+        return 0.5f;
     }
 
     @Override
