+++ common/src/main/java/com/fossil/fossil/entity/prehistoric/Liopleurodon.java
@@ -101,7 +101,7 @@ public class Liopleurodon extends PrehistoricSwimming {
     }
 
     @Override
-    public double swimSpeed() {
+    public float swimSpeed() {
         return 6;
     }
