+++ common/src/main/java/com/fossil/fossil/entity/prehistoric/Megalodon.java
@@ -82,7 +82,7 @@ public class Megalodon extends PrehistoricSwimming {
     }
 
     @Override
-    public double swimSpeed() {
+    public float swimSpeed() {
         return 1;
     }
