+++ common/src/main/java/com/fossil/fossil/entity/prehistoric/Megalograptus.java
@@ -79,8 +79,8 @@ public class Megalograptus extends PrehistoricSwimming {
     }
 
     @Override
-    public double swimSpeed() {
-        return 0.5;
+    public float swimSpeed() {
+        return 0.5f;
     }
 
     @Override
