+++ common/src/main/java/com/fossil/fossil/entity/prehistoric/Plesiosaurus.java
@@ -76,7 +76,7 @@ public class Plesiosaurus extends PrehistoricSwimming {
     }
 
     @Override
-    public double swimSpeed() {
+    public float swimSpeed() {
         return 1;
     }
