+++ common/src/main/java/com/fossil/fossil/entity/prehistoric/Sarcosuchus.java
@@ -80,7 +80,7 @@ public class Sarcosuchus extends PrehistoricSwimming implements PrehistoricScary
     }
 
     @Override
-    public double swimSpeed() {
+    public float swimSpeed() {
         return 1;
     }
