+++ common/src/main/java/com/fossil/fossil/entity/prehistoric/Tiktaalik.java
@@ -78,8 +78,8 @@ public class Tiktaalik extends PrehistoricSwimming {
     }
 
     @Override
-    public double swimSpeed() {
-        return 0.5;
+    public float swimSpeed() {
+        return 0.5f;
     }
 
     @Override
